% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculates GTI scores for input subjects and returns the results}
\usage{
calculate_score(
  input_subjects,
  baseline = "previous",
  include = list(),
  output_format = NULL,
  algorithm = "agti",
  unit = 0,
  fill_missing = FALSE
)
}
\arguments{
\item{input_subjects}{A DataFrame or list of lists representing the input subjects.}

\item{baseline}{A string indicating the baseline visit to use for calculating changes. Default is "previous".}

\item{include}{A list of strings indicating which measures to include in the output. Default is an empty list, which includes all measures.}

\item{output_format}{Either 'json' or 'df'. By default, it returns the same as input_subjects type.}

\item{algorithm}{Indicates which algorithm to use. Default is 'agti' (adults).}

\item{unit}{Indicates whether to use metric (0) or imperial(1) units. Default is 0 (metric).}

\item{fill_missing}{Indicates whether to fill missing values or not. Default is FALSE.}
}
\value{
A JSON-formatted string representing the output scores for each measure and subject.
}
\description{
Calculates GTI scores for input subjects and returns the results
}
\examples{
\dontrun{
calculate_score(subjects, baseline = "previous", algorithm = "agti")
}
}
